sequencer works! barely has bugs i believe
