+++ client/routes.js
@@ -2,9 +2,8 @@ import React, { Component } from "react";
 import { connect } from "react-redux";
 import { withRouter, Route, Switch, Redirect } from "react-router-dom";
 import Home from "./components/Home";
-import Sequencer from "./components/Sequencer";
 import { WaitingRoom } from "./components/WaitingRoom";
-import { SeqAttempt } from "./components/SequencerAttempt";
+import { Sequencer } from "./components/Sequencer";
 
 /**
  * COMPONENT
@@ -15,7 +14,7 @@ class Routes extends Component {
       <div>
         <Switch>
           <Route exact path="/" component={Home} />
-          <Route exact path="/sequencer" component={SeqAttempt} />
+          <Route exact path="/sequencer" component={Sequencer} />
           <Route path="/waiting/:roomId" component={WaitingRoom} />
           {/* <Redirect to="/" /> */}
         </Switch>
